requirements.txt
allure-pytest
allure-python-commons
msedge-selenium-tools
pytest
requests
selenium
webdriver-manager
